+++ packages/nesd/lib/nes/isolate/nes_worker.dart
@@ -165,6 +165,8 @@ class NesWorker {
         _audioOutput?.lowPassFilter = command.enabled;
       case SetSwapDutyCyclesCommand():
         _nes?.apu.swapDutyCycles = command.enabled;
+      case SetMixerCommand():
+        _nes?.apu.mixer = command.mixer;
       case SetPaletteCommand():
         _nes?.ppu.systemPalette = command.palette;
       case StartPcmDumpCommand():